consolitate tests
