+++ internal/api/handlers/ai_models_test.go
@@ -19,6 +19,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/arduino/arduino-app-cli/internal/api/models"
+	"github.com/arduino/arduino-app-cli/internal/orchestrator/bricks"
 	"github.com/arduino/arduino-app-cli/internal/orchestrator/config"
 	"github.com/arduino/arduino-app-cli/internal/orchestrator/modelsindex"
 	"github.com/arduino/arduino-app-cli/internal/platform"
@@ -107,9 +108,6 @@ func TestDownloadStream(t *testing.T) {
 		assert.Equal(t, sseLog{Message: "download complete"}, sse.events[0].Data)
 	})
 
-}
-
-func TestDownloadStreamSendError(t *testing.T) {
 	t.Run("a full models directory carries its own code", func(t *testing.T) {
 		// Wrapped, as Download reports it: the client shows a different message for a
 		// disk that is full than for a download that broke.
@@ -295,5 +293,37 @@ func TestHandlerModelByID(t *testing.T) {
 		HandlerModelByID(testModelsIndex(t))(rec, req)
 
 		assert.Equal(t, http.StatusNotFound, rec.Code)
+	})
+}
+
+// TestDecodeRequestModel covers the other door a model id comes in by: a brick request
+// names one in its body, in the same base64url form the path takes.
+func TestDecodeRequestModel(t *testing.T) {
+	t.Run("an encoded id becomes the plain one", func(t *testing.T) {
+		encoded := models.EncodeModelID("llamacpp:owner/repo/file")
+		req := bricks.BrickCreateUpdateRequest{Model: &encoded}
+
+		require.NoError(t, decodeRequestModel(&req))
+		assert.Equal(t, "llamacpp:owner/repo/file", *req.Model)
+	})
+
+	t.Run("a request naming no model is left alone", func(t *testing.T) {
+		req := bricks.BrickCreateUpdateRequest{}
+		require.NoError(t, decodeRequestModel(&req))
+		assert.Nil(t, req.Model)
+
+		empty := ""
+		req = bricks.BrickCreateUpdateRequest{Model: &empty}
+		require.NoError(t, decodeRequestModel(&req))
+		assert.Equal(t, "", *req.Model)
+	})
+
+	t.Run("an id that is not base64url is refused", func(t *testing.T) {
+		plain := "llamacpp:owner/repo/file"
+		req := bricks.BrickCreateUpdateRequest{Model: &plain}
+
+		err := decodeRequestModel(&req)
+		require.Error(t, err)
+		assert.Contains(t, err.Error(), "base64url")
 	})
 }
